starting culling tiles off the screen, doesnt work if moving right/down

+++ WorldConstants.h
@@ -1,6 +1,6 @@
 #ifndef WORLDCONSTANTS_H
 #define WORLDCONSTANTS_H
 
-const float TILE_SIZE = 50.f;
+const int TILE_SIZE = 50.f;
 
 #endif
